requirements.txt
# Python dependencies for Supabase integration
supabase==2.3.4
python-dotenv==1.0.0
